requirements.txt
fastapi
uvicorn
python-multipart
pydantic
pydantic-settings
google-generativeai
pdfplumber
python-dotenv
python-docx
PyPDF2
jinja2
aiofiles
